# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Category:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      logoHashedID:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
  models.CategoryWithStats:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      logoHashedID:
        type: string
      name:
        type: string
      operations_count:
        type: integer
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
...
  models.ErrorCode:
    enum:
    - INVALID_REQUEST
    - MISSING_FIELDS
    - INVALID_EMAIL
    - INVALID_PASSWORD
    - INVALID_LOGIN
    - USER_EXISTS
    - EMAIL_EXISTS
    - LOGIN_EXISTS
    - WEAK_PASSWORD
    - USER_NOT_FOUND
    - INVALID_CREDENTIALS
    - ACCOUNT_LOCKED
    - TOKEN_EXPIRED
    - TOKEN_INVALID
    - TOKEN_MISSING
    - UNAUTHORIZED
    - FORBIDDEN
    - ACCESS_DENIED
    - RESOURCE_NOT_FOUND
    - RESOURCE_EXISTS
    - RESOURCE_CONFLICT
    - BUDGET_NOT_FOUND
    - ACCOUNT_NOT_FOUND
    - OPERATION_NOT_FOUND
    - INVALID_AMOUNT
    - INVALID_CURRENCY
    - INVALID_DATE
    - INVALID_PERIOD
    - INTERNAL_ERROR
    - SERVICE_UNAVAILABLE
    - DATABASE_ERROR
    type: string
    x-enum-varnames:
    - ErrCodeInvalidRequest
    - ErrCodeMissingFields
    - ErrCodeInvalidEmail
    - ErrCodeInvalidPassword
    - ErrCodeInvalidLogin
    - ErrCodeUserExists
    - ErrCodeEmailExists
    - ErrCodeLoginExists
    - ErrCodeWeakPassword
    - ErrCodeUserNotFound
    - ErrCodeInvalidCredentials
    - ErrCodeAccountLocked
    - ErrCodeTokenExpired
    - ErrCodeTokenInvalid
    - ErrCodeTokenMissing
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeAccessDenied
    - ErrCodeResourceNotFound
    - ErrCodeResourceExists
    - ErrCodeResourceConflict
    - ErrCodeBudgetNotFound
    - ErrCodeAccountNotFound
    - ErrCodeTransactionNotFound
    - ErrCodeInvalidAmount
    - ErrCodeInvalidCurrency
    - ErrCodeInvalidDate
    - ErrCodeInvalidPeriod
    - ErrCodeInternalError
    - ErrCodeServiceUnavailable
    - ErrCodeDatabaseError
  models.ErrorResponse:
    properties:
      code:
        $ref: '#/definitions/models.ErrorCode'
      details:
        type: string
      error:
        type: string
      field:
        type: string
      timestamp:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для управления личными финансами
  title: VKarmane API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Возвращает список всех категорий пользователя с количеством операций
      produces:
      - application/json
      responses:
        "200":
          description: Список категорий пользователя
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Требуется аутентификация (UNAUTHORIZED, TOKEN_MISSING, TOKEN_INVALID,
            TOKEN_EXPIRED)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR, DATABASE_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение списка категорий пользователя
      tags:
      - categories
    post:
      consumes:
      - multipart/form-data
      description: Создает новую категорию для пользователя. Поддерживает multipart/form-data
        с опциональным полем image для загрузки картинки категории
      parameters:
      - description: Название категории
        in: formData
        name: name
        required: true
        type: string
      - description: Описание категории
        in: formData
        name: description
        type: string
      - description: Картинка категории (опционально)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Созданная категория
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Некорректные данные (VALIDATION_ERROR, INVALID_INPUT)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Требуется аутентификация (UNAUTHORIZED, TOKEN_MISSING, TOKEN_INVALID,
            TOKEN_EXPIRED)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR, DATABASE_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создание новой категории
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Удаляет категорию пользователя
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Категория успешно удалена
        "401":
          description: Требуется аутентификация (UNAUTHORIZED, TOKEN_MISSING, TOKEN_INVALID,
            TOKEN_EXPIRED)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Категория не найдена (RESOURCE_NOT_FOUND)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR, DATABASE_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удаление категории
      tags:
      - categories
    get:
      description: Возвращает информацию о категории по её ID с количеством операций
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о категории
          schema:
            $ref: '#/definitions/models.CategoryWithStats'
        "401":
          description: Требуется аутентификация (UNAUTHORIZED, TOKEN_MISSING, TOKEN_INVALID,
            TOKEN_EXPIRED)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Категория не найдена (RESOURCE_NOT_FOUND)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR, DATABASE_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение категории по ID
      tags:
      - categories
    put:
      consumes:
      - multipart/form-data
      description: Обновляет информацию о категории. Поддерживает multipart/form-data
        с опциональным полем image для загрузки картинки категории
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      - description: Название категории
        in: formData
        name: name
        type: string
      - description: Описание категории
        in: formData
        name: description
        type: string
      - description: Картинка категории (опционально)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная категория
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Некорректные данные (VALIDATION_ERROR, INVALID_INPUT)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Требуется аутентификация (UNAUTHORIZED, TOKEN_MISSING, TOKEN_INVALID,
            TOKEN_EXPIRED)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Категория не найдена (RESOURCE_NOT_FOUND)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR, DATABASE_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновление категории
      tags:
      - categories
...
securityDefinitions:
  ApiKeyAuth:
    description: JWT токен в HTTP-only cookie
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"